move ui components to there own files and improve package structure
